fix: clean up responsibilities of typing tool and example
