+++ common/changes/@kickstartds/jsonschema2types/fix-type-generation-responsibilities_2023-09-24-20-29.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@kickstartds/jsonschema2types",
+      "comment": "don't write files, return record of generated types",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@kickstartds/jsonschema2types"
+}
