+++ common/changes/@kickstartds/types-example/fix-type-generation-responsibilities_2023-09-24-20-29.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@kickstartds/types-example",
+      "comment": "write files in example code now",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@kickstartds/types-example"
+}
