+++ examples/types/package.json
@@ -34,6 +34,7 @@
     "eslint-plugin-json-files": "~2.2.0",
     "import-meta-resolve": "~3.0.0",
     "json-schema-typed": "~8.0.1",
+    "pascal-case": "^3.1.2",
     "prettier": "^2.8.8",
     "typescript": "^5.0.4"
   },
