+++ examples/types/src/index.ts
@@ -1,15 +1,22 @@
+import { mkdirSync, writeFileSync } from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import {
   getCustomSchemaIds,
+  getSchemaModule,
+  getSchemaName,
   getSchemaRegistry,
   getUniqueSchemaIds,
+  isLayering,
+  layeredSchemaId,
   processSchemaGlob,
   shouldLayer
 } from '@kickstartds/jsonschema-utils';
 import { createTypes } from '@kickstartds/jsonschema2types';
 import { resolve } from 'import-meta-resolve';
+import { type JSONSchema } from 'json-schema-typed/draft-07';
+import { pascalCase } from 'pascal-case';
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
 (async () => {
@@ -18,8 +25,6 @@ import { resolve } from 'import-meta-resolve';
   );
   const customGlob = `${packagePath}/(dist|cms)/**/*.(schema|definitions).json`;
 
-  // get shared ajv instance, pre-process schemas and get full
-  // set of unique schemas. precondition for the following conversions
   const ajv = getSchemaRegistry();
   const schemaIds = await processSchemaGlob(customGlob, ajv, false);
   const kdsSchemaIds = schemaIds.filter((schemaId) => schemaId.includes('schema.kickstartds.com'));
@@ -32,5 +37,26 @@ import { resolve } from 'import-meta-resolve';
     kdsSchemaIds.some((kdsSchemaId) => shouldLayer(schemaId, kdsSchemaId))
   );
 
-  await createTypes([...unlayeredSchemaIds, ...layeredSchemaIds], kdsSchemaIds, ajv);
+  const layeredTypes = await createTypes([...unlayeredSchemaIds, ...layeredSchemaIds], ajv);
+
+  mkdirSync('dist', { recursive: true });
+
+  for (const schemaId of Object.keys(layeredTypes)) {
+    const schema = ajv.getSchema(schemaId)?.schema as JSONSchema.Interface;
+
+    if (!schema) throw new Error("Can't find schema for layered type");
+    if (!schema.$id) throw new Error('Found schema without $id property');
+
+    const layeredId = isLayering(schema.$id, kdsSchemaIds)
+      ? layeredSchemaId(schema.$id, kdsSchemaIds)
+      : schema.$id;
+
+    writeFileSync(
+      `dist/${pascalCase(getSchemaName(layeredId))}Props.ts`,
+      `declare module "@kickstartds/${getSchemaModule(layeredId)}/lib/${getSchemaName(layeredId)}/typing" {
+${layeredTypes[schemaId]}
+}
+        `
+    );
+  }
 })();
